+++ prepline_general/api/general.py
@@ -3,7 +3,9 @@
 # DO NOT MODIFY DIRECTLY
 #####################################################################
 
+import io
 import os
+import gzip
 import mimetypes
 from typing import List, Union
 from fastapi import (
@@ -19,6 +21,7 @@ from fastapi import (
 from fastapi.responses import PlainTextResponse
 import json
 from fastapi.responses import StreamingResponse
+from starlette.datastructures import Headers
 from starlette.types import Send
 from base64 import b64encode
 from typing import Optional, Mapping, Iterator, Tuple
@@ -42,19 +45,64 @@ def is_expected_response_type(media_type, response_type):
 
 
 # pipeline-api
-def pipeline_api(file, filename="", response_type="application/json"):
-    # NOTE(robinson) - This is a hacky solution due to
-    # limitations in the SpooledTemporaryFile wrapper.
-    # Specifically, it does't have a `seekable` attribute,
-    # which is required for .pptx and .docx. See below
-    # the link below
-    # ref: https://stackoverflow.com/questions/47160211
-    # /why-doesnt-tempfile-spooledtemporaryfile-implement-readable-writable-seekable
-    with tempfile.TemporaryDirectory() as tmpdir:
-        _filename = os.path.join(tmpdir, filename.split("/")[-1])
-        with open(_filename, "wb") as f:
-            f.write(file.read())
-        elements = partition(filename=_filename)
+
+
+DEFAULT_MIMETYPES = (
+    "application/pdf,application/msword,image/jpeg,image/png,text/markdown,"
+    "text/x-markdown,text/html,"
+    "application/vnd.openxmlformats-officedocument.wordprocessingml.document,"
+    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet,"
+    "application/vnd.ms-excel,application/vnd.openxmlformats-officedocument."
+    "presentationml.presentation,"
+    "application/vnd.ms-powerpoint,application/xml"
+    "text/html,message/rfc822,text/plain,image/png,"
+)
+
+if not os.environ.get("UNSTRUCTURED_ALLOWED_MIMETYPES", None):
+    os.environ["UNSTRUCTURED_ALLOWED_MIMETYPES"] = DEFAULT_MIMETYPES
+
+
+def pipeline_api(
+    file,
+    filename="",
+    m_strategy=[],
+    file_content_type=None,
+    response_type="application/json",
+):
+    strategy = (m_strategy[0] if len(m_strategy) else "fast").lower()
+    if strategy not in ["fast", "hi_res"]:
+        raise HTTPException(
+            status_code=400,
+            detail=f"Invalid strategy: {strategy}. Must be one of ['fast', 'hi_res']",
+        )
+    if filename.endswith((".docx", ".pptx")):
+        # NOTE(robinson) - This is a hacky solution due to
+        # limitations in the SpooledTemporaryFile wrapper.
+        # Specifically, it doesn't have a `seekable` attribute,
+        # which is required for .pptx and .docx. See below
+        # the link below
+        # ref: https://stackoverflow.com/questions/47160211
+        # /why-doesnt-tempfile-spooledtemporaryfile-implement-readable-writable-seekable
+        with tempfile.TemporaryDirectory() as tmpdir:
+            _filename = os.path.join(tmpdir, filename.split("/")[-1])
+            with open(_filename, "wb") as f:
+                f.write(file.read())
+            elements = partition(filename=_filename, strategy=strategy)
+    else:
+        try:
+            elements = partition(
+                file=file,
+                file_filename=filename,
+                content_type=file_content_type,
+                strategy=strategy,
+            )
+        except ValueError as e:
+            if "Invalid file" in e.args[0]:
+                raise HTTPException(
+                    status_code=400,
+                    detail=f"{file_content_type} not currently supported",
+                )
+            raise e
 
     # Due to the above, elements have an ugly temp filename in their metadata
     # For now, replace this with the basename
@@ -72,7 +120,7 @@ def get_validated_mimetype(file):
     return HTTP 400 for an invalid type.
     """
     content_type = file.content_type
-    if content_type == "application/octet-stream":
+    if not content_type or content_type == "application/octet-stream":
         content_type = mimetypes.guess_type(str(file.filename))[0]
 
         # Markdown mimetype is too new for the library - just hardcode that one in for now
@@ -85,7 +133,11 @@ def get_validated_mimetype(file):
 
         if content_type not in allowed_mimetypes:
             raise HTTPException(
-                status_code=400, detail=f"File type not supported: {file.filename}"
+                status_code=400,
+                detail=(
+                    f"Unable to process {file.filename}: "
+                    f"File type {content_type} is not supported."
+                ),
             )
 
     return content_type
@@ -149,13 +201,30 @@ class MultipartMixedResponse(StreamingResponse):
         await send({"type": "http.response.body", "body": b"", "more_body": False})
 
 
+def ungz_file(file: UploadFile) -> UploadFile:
+    filename = str(file.filename) if file.filename else ""
+    gzip_file = gzip.open(file.file)
+    return UploadFile(
+        file=io.BytesIO(gzip_file.read()),
+        size=len(gzip_file.read()),
+        filename=filename[:-3] if len(filename) > 3 else "",
+        headers=Headers({"content-type": str(mimetypes.guess_type(filename)[0])}),
+    )
+
+
 @router.post("/general/v0/general")
-@router.post("/general/v0.0.8/general")
+@router.post("/general/v0.0.11/general")
 def pipeline_1(
     request: Request,
     files: Union[List[UploadFile], None] = File(default=None),
     output_format: Union[str, None] = Form(default=None),
+    strategy: List[str] = Form(default=[]),
 ):
+    if files:
+        for file_index in range(len(files)):
+            if files[file_index].content_type == "application/gzip":
+                files[file_index] = ungz_file(files[file_index])
+
     content_type = request.headers.get("Accept")
 
     default_response_type = output_format or "application/json"
@@ -181,14 +250,16 @@ def pipeline_1(
 
             def response_generator(is_multipart):
                 for file in files:
-                    _ = get_validated_mimetype(file)
+                    file_content_type = get_validated_mimetype(file)
 
                     _file = file.file
 
                     response = pipeline_api(
                         _file,
+                        m_strategy=strategy,
                         response_type=media_type,
                         filename=file.filename,
+                        file_content_type=file_content_type,
                     )
                     if is_multipart:
                         if type(response) not in [str, bytes]:
@@ -205,12 +276,14 @@ def pipeline_1(
             file = files[0]
             _file = file.file
 
-            _ = get_validated_mimetype(file)
+            file_content_type = get_validated_mimetype(file)
 
             response = pipeline_api(
                 _file,
+                m_strategy=strategy,
                 response_type=media_type,
                 filename=file.filename,
+                file_content_type=file_content_type,
             )
 
             if is_expected_response_type(media_type, type(response)):
@@ -235,11 +308,6 @@ def pipeline_1(
             content='Request parameter "files" is required.\n',
             status_code=status.HTTP_400_BAD_REQUEST,
         )
-
-
-@app.get("/healthcheck", status_code=status.HTTP_200_OK)
-def healthcheck(request: Request):
-    return {"healthcheck": "HEALTHCHECK STATUS: EVERYTHING OK!"}
 
 
 app.include_router(router)